Add check if the path exists before removing it.

+++ pyvista/examples/downloads.py
@@ -84,7 +84,8 @@ def delete_downloads():
 
     """
     _check_examples_path()
-    shutil.rmtree(pyvista.EXAMPLES_PATH)
+    if os.path.exists(pyvista.EXAMPLES_PATH):
+        shutil.rmtree(pyvista.EXAMPLES_PATH)
     os.makedirs(pyvista.EXAMPLES_PATH)
     return True
